+++ imgui-binding/src/main/java/imgui/ImFontConfig.java
@@ -42,7 +42,7 @@ public final class ImFontConfig implements ImDestroyable {
 
     private native long nCreate(); /*
         ImFontConfig* imFontConfig = new ImFontConfig();
-        return (long)(intptr_t)imFontConfig;
+        return (intptr_t)imFontConfig;
     */
 
     private native void nDestroy(long ptr); /*
